backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-restful>=0.3.10",
    "flask-security-too>=5.6.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "weasyprint>=65.1",
]
